+++ bugsnag-android-core/src/androidTest/java/com/bugsnag/android/AppDataTest.java
@@ -10,7 +10,8 @@ import static org.mockito.Mockito.when;
 
 import android.content.Context;
 import android.content.pm.PackageManager;
-import android.support.test.InstrumentationRegistry;
+
+import androidx.test.core.app.ApplicationProvider;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -43,7 +44,7 @@ public class AppDataTest {
         when(sessionTracker.isInForeground()).thenReturn(true);
         when(sessionTracker.getDurationInForegroundMs(anyLong())).thenReturn(500L);
 
-        Context context = InstrumentationRegistry.getContext();
+        Context context = ApplicationProvider.getApplicationContext();
         PackageManager packageManager = context.getPackageManager();
         Configuration config = new Configuration("api-key");
         AppData obj = new AppData(context, packageManager, config, sessionTracker);