+++ bugsnag-android-core/src/androidTest/java/com/bugsnag/android/ClientNotifyTest.java
@@ -2,20 +2,17 @@ package com.bugsnag.android;
 
 import static org.junit.Assert.assertEquals;
 
-import android.support.annotation.NonNull;
-import android.support.test.filters.SmallTest;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.annotation.NonNull;
+import androidx.test.filters.SmallTest;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
-@RunWith(AndroidJUnit4.class)
 @SmallTest
 public class ClientNotifyTest {