+++ bugsnag-android-core/src/androidTest/java/com/bugsnag/android/ExceptionHandlerTest.java
@@ -5,16 +5,14 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.filters.SmallTest;
-import android.support.test.runner.AndroidJUnit4;
+
+import androidx.test.core.app.ApplicationProvider;
+import androidx.test.filters.SmallTest;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-@RunWith(AndroidJUnit4.class)
 @SmallTest
 public class ExceptionHandlerTest {
 
@@ -29,7 +27,7 @@ public class ExceptionHandlerTest {
      */
     @Before
     public void setUp() throws Exception {
-        context = InstrumentationRegistry.getContext();
+        context = ApplicationProvider.getApplicationContext();
         // Start in a clean state, since we've created clients before in tests
         Thread.setDefaultUncaughtExceptionHandler(null);
         client = new Client(context, "api-key");