+++ bugsnag-android-core/src/main/java/com/bugsnag/android/BugsnagException.java
@@ -1,6 +1,6 @@
 package com.bugsnag.android;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 /**
  * Used to store information about an exception that was not provided with an exception object