+++ bugsnag-android-core/src/main/java/com/bugsnag/android/Callback.java
@@ -1,6 +1,6 @@
 package com.bugsnag.android;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 /**
  * A callback to be run before an individual report is sent to Bugsnag.