+++ bugsnag-android-core/src/main/java/com/bugsnag/android/HandledState.java
@@ -1,9 +1,9 @@
 package com.bugsnag.android;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.StringDef;
 import android.text.TextUtils;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.StringDef;
 
 import java.io.IOException;
 import java.lang.annotation.Retention;