+++ bugsnag-android-core/src/main/java/com/bugsnag/android/SessionTrackingPayload.java
@@ -1,6 +1,6 @@
 package com.bugsnag.android;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.io.File;
 import java.io.IOException;